Divide classes into packages.
