+++ src/main/java/duke/command/ByeCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.helpers.Storage;
+import duke.helpers.Ui;
+import duke.task.TaskList;
+
 public class ByeCommand extends Command {
