+++ src/main/java/duke/command/CommandType.java
@@ -1,3 +1,5 @@
+package duke.command;
+
 /**
  * Enums of all command type.
  */
