+++ src/main/java/duke/command/PrintCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.helpers.Storage;
+import duke.helpers.Ui;
+import duke.task.TaskList;
+
 public class PrintCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
