+++ src/main/java/duke/helpers/FileManaging.java
@@ -1,8 +1,13 @@
+package duke.helpers;
+
+import duke.exception.DukeException;
+import duke.task.Task;
+import duke.task.TaskList;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.util.List;
 import java.util.Scanner;
 
 public class FileManaging {
@@ -19,7 +24,7 @@ public class FileManaging {
         Scanner s = new Scanner(f);
 
         while (s.hasNext()) {
-            // Create a Task object and store into local array.
+            // Create a duke.task.Task object and store into local array.
             String line = s.nextLine();
             String[] arr = line.split(" , ");
 
@@ -36,7 +41,7 @@ public class FileManaging {
                 Task curr = Task.of(arr[0], Integer.parseInt(arr[1]) == 1, arr[2], from, to);
                 storage.add(curr);
             } else {
-                throw new DukeException("Storage Format Issue");
+                throw new DukeException("duke.helpers.Storage Format Issue");
             }
         }
     }
