+++ src/main/java/duke/helpers/Storage.java
@@ -1,3 +1,9 @@
+package duke.helpers;
+
+import duke.command.CommandType;
+import duke.exception.DukeException;
+import duke.task.TaskList;
+
 import java.io.FileNotFoundException;
 
 public class Storage {
