+++ src/main/java/duke/task/Duke.java
@@ -1,11 +1,11 @@
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Scanner;
+package duke.task;
+
+import duke.command.Command;
+import duke.command.CommandType;
+import duke.helpers.Parser;
+import duke.helpers.Storage;
+import duke.helpers.Ui;
+
 public class Duke {
     private Ui ui;
     private TaskList tasks;
