+++ storage.txt
@@ -1,5 +1,5 @@
 todo , 1 , todo1
-deadline , 1 , deadline1 , 2024-01-01 , 06:30
+deadline , 0 , deadline1 , 2024-01-01 , 06:30
 event , 1 , event1 , 2024-01-01 , 06:30 , 2024-01-02 , 06:30
 todo , 0 , study
-deadline , 0 , new job , 2024-02-02 , 11:30
+deadline , 1 , new job , 2024-02-02 , 11:30
